lab_pneumonia/enhanced_deployment/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
numpy==1.23.5
Pillow==9.5.0
keras==2.12.0
tensorflow==2.12.0

